Refactoring to prepare for hspec-wai
